Simplify Stellar::Price rational approximation logic
Use stdlib `rationalize` method instead of the custom
implementation in `Util::ContinuedFraction`.

Primary motivation for this change was that `Util::ContinuedFraction`
is extremely slow when used with BigDecimal on JRuby, so that a single
test from our suite was unable to complete and timing out after 10
minutes on CI.

It's less code to maintain also.